Rename Main to BookkeeperStarter for Improved Clarity and Context
